+++ src/App.jsx
@@ -1,7 +1,7 @@
 import { Routes, Route } from 'react-router-dom';
 import Header from './Components/Header';
 import Profile from './Components/Profile';
-import DisplayRockets from './Components/DisplayRockets';
+import DisplayRockets from './Components/Rockets/DisplayRockets';
 import DisplayDragons from './Components/Dragons/DisplayDragons';
 import MissionsPage from './Components/Missions/missionsPage';
